+++ src/inspect_ai/scorer/_model.py
@@ -18,7 +18,7 @@ from inspect_ai.model._model_output import ModelOutput
 from inspect_ai.solver._task_state import TaskState
 from inspect_ai.util import resource
 
-from ._metric import INCORRECT, Score
+from ._metric import NOANSWER, Score
 from ._metrics import accuracy, stderr
 from ._multi import multi_scorer
 from ._scorer import Scorer, scorer
@@ -223,7 +223,7 @@ def _model_graded_qa_single(
             )
         else:
             return Score(
-                value=INCORRECT,
+                value=NOANSWER,
                 answer=state.output.completion,
                 explanation="Grade not found in model output: "
                 + f"{result.completion}",